requirements.txt
fastapi
uvicorn
pydantic
streamlit
requests
google-cloud-vision
google-generativeai
PyMuPDF
Pillow
pandas
python-dotenv
